--- poseidon/beaglebone/agent/extern/utils.js
@@ -1,1 +1,0 @@
-module.exports = require('../../../../utils/javascript/lib')
