+++ poseidon/beaglebone/agent/utils.js
@@ -1,0 +1,1 @@
+module.exports = require('../../../utils/javascript/lib')
