+++ poseidon/beaglebone/system/run.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 set -e
-# build
 mkdir -p build
 cd build
-cmake .. -Dtests=OFF -Dcoverage=OFF
+cmake ..
 make
+./system
