--- poseidon/beaglebone/system/src/test.cpp
@@ -1,1 +1,0 @@
-#include <test.hpp>
