--- poseidon/beaglebone/system/test/main.cpp
@@ -1,2 +1,0 @@
-#define CATCH_CONFIG_MAIN  // This tells Catch to provide a main() - only do this in one cpp file
-#include "catch2/catch.hpp"
