--- poseidon/raspberryPI/include/test.hpp
@@ -1,3 +1,0 @@
-int test(){
-  return 0;
-}
