+++ poseidon/raspberryPI/run.sh
@@ -3,5 +3,6 @@ set -e
 # build
 mkdir -p build
 cd build
-cmake .. -Dtests=OFF -Dcoverage=OFF
+cmake ..
 make
+./vision
