--- poseidon/raspberryPI/test.sh
@@ -1,8 +1,0 @@
-#!/bin/bash
-set -e
-# run tests
-mkdir -p build
-cd build
-cmake .. -Dtests=ON -Dcoverage=OFF
-make
-./raspberryPI_tests
