+++ utils/cpp/src/Vision/Streameye.cpp
@@ -1,79 +1,35 @@
 #ifdef ENABLE_OPEN_CV
 
-#include <Vision/StreamEye.h>
-int ofile = -1;
+#include <Vision/Streameye.hpp>
 
-StreamEye::StreamEye() {
-  std::cout << "Setting up StreamEye" << std::endl;
-  ofile = open("log.txt",O_WRONLY|O_CREAT);
-  if ( pipe(m_fdIn) < 0  ||  pipe(m_fdOut) < 0 ){
-  	/* FATAL: cannot create pipe */
-  	/* close readpipe[0] & [1] if necessary */
-    std::cout << "Error pipe" << std::endl;
+StreamEye::StreamEye(unsigned int port, int quality)
+  : streamEyeApp(APP_NAME,
+      {"-q", "-p", std::to_string(port), "-s", SEPERATOR}
+    ),
+    compressionConfig({CV_IMWRITE_JPEG_QUALITY, quality}),
+    timeR(0), timeG(0), timeB(0),
+    showTime(false) {}
+
+void StreamEye::serveFrame(cv::Mat frame) {
+  if(this->showTime){
+    std::time_t t = std::chrono::system_clock::to_time_t(std::chrono::system_clock::now());
+    cv::putText(frame, std::ctime(&t), cv::Point(30, frame.rows - 30), cv::FONT_HERSHEY_COMPLEX_SMALL, 1, cv::Scalar(this->timeB,this->timeG,this->timeR), 1, CV_AA);
   }
-  std::cout << m_fdIn[0] << ":" << m_fdIn[1] << ":" << m_fdOut[0] << ":" << m_fdOut[1] << ":" << std::endl;
-  auto pid = fork();
-	if(pid < 0){
-		std::perror("Error");
-	}else if(pid==0){
-		this->childProcess();
-	}else{
-    m_PID = pid;
-    close(m_fdIn[0]);
-    close(m_fdOut[1]);
-    close(ofile);
-    char nChar;
-    // while(read(m_fdOut[0], &nChar, 1)==1) {
-    //   std::cout << "Char:" << nChar << std::endl;
-    //   write(STDOUT_FILENO,&nChar, 1);
-    // }
-	}
+  std::vector<unsigned char> image;
+  cv::imencode(".jpg", frame, image, this->compressionConfig);
+  this->streamEyeApp.write((unsigned char*)SEPERATOR, sizeof(SEPERATOR)-1);
+  this->streamEyeApp.write((unsigned char*)&image[0], sizeof(unsigned char)*image.size());
 }
 
-void StreamEye::childProcess(){
-  // dup2(m_fdIn[0],STDIN_FILENO);
-  // dup2(m_fdOut[1],STDOUT_FILENO);
-  close(m_fdIn[1]);
-  close(m_fdOut[0]);
-  // if(dup2(m_fdOut[1],STDOUT_FILENO) == -1) {
-  //   exit(errno);
-  // }
-
-  // if(dup2(ofile,STDOUT_FILENO) == -1) {
-  //    exit(errno);
-  // }
-  // if(dup2(STDOUT_FILENO,STDERR_FILENO) == -1) {
-  //    exit(errno);
-  // }
-  if(dup2(m_fdIn[0],STDIN_FILENO) == -1) {
-    exit(errno);
-  }
-  close(ofile);
-  // close(m_fdIn[0]);
-  // close(m_fdOut[1]);
-  std::cout << "Dup2 done" << std::endl;
-  char* const  cmd[] = { "streameye","-q","-p","8080","-s",SEPERATOR,NULL };
-  //char* const  cmd[] = { "echo"/*,"-q"*/,"-p","8080","-s",SEPERATOR,NULL };
-  int result = execvp(cmd[0], &cmd[0]);
-  std::perror(cmd[0]);
-  exit(result);
+void StreamEye::overlayTime(int r, int g, int b){
+  this->timeR = r;
+  this->timeG = g;
+  this->timeB = b;
+  this->showTime = true;
 }
 
-void StreamEye::writeJpeg(std::vector<unsigned char> image) {
-  write(m_fdIn[1], SEPERATOR, sizeof(SEPERATOR)-1);
-  write(m_fdIn[1], (char *) &image[0], sizeof(unsigned char)*image.size());
-}
-
-StreamEye::~StreamEye() {
-  int status;
-  //close(m_fdIn[0]);
-  close(m_fdIn[1]);
-  close(m_fdOut[0]);
-  close(ofile);
-  //close(m_fdOut[1]);
-  std::cout << "Closed Pipes" << std::endl;
-  waitpid(m_PID,&status,0);
-  std::cout << "Process finished" << std::endl;
+void StreamEye::stopOverlay(){
+  this->showTime = false;
 }
 
 #endif
